+++ src/nvim/ex_docmd.c
@@ -38,6 +38,7 @@
 #include "nvim/getchar.h"
 #include "nvim/globals.h"
 #include "nvim/hardcopy.h"
+#include "nvim/highlight_group.h"
 #include "nvim/if_cscope.h"
 #include "nvim/input.h"
 #include "nvim/keymap.h"
