+++ src/nvim/memory.c
@@ -13,6 +13,7 @@
 #include "nvim/decoration_provider.h"
 #include "nvim/eval.h"
 #include "nvim/highlight.h"
+#include "nvim/highlight_group.h"
 #include "nvim/lua/executor.h"
 #include "nvim/memfile.h"
 #include "nvim/memory.h"
